Move partition projection to io.trino.plugin.hive.projection package
